requirements.txt
# Core Django
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3

# Django REST Framework
djangorestframework==3.16.1
django-filter==24.3
markdown==3.7

# Database
psycopg2-binary==2.9.9  # Para PostgreSQL en producción

# Caching
redis==5.1.1
django-redis==5.4.0

# Development & Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
faker==30.3.0

# Code Quality
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Performance & Monitoring
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Security
django-cors-headers==4.4.0

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Environment
python-decouple==3.8
